Request to update DZone link
In the 'What is time-series data, and why are we building a time-series database (TSDB)?' blog, at the end of the 'QuestDB design and performance' section (https://questdb.io/blog/2020/11/26/why-timeseries-data/#questdb-design-and-performance), the 'on DZone' link (https://dzone.com/articles/we-put-a-sql-database-on-the-internet) does not display the article. 

Within DZone too, I was not able to locate the article using the 'Search' option. So, I emailed support@dzone.com and asked them for the correct link, if available. I will update the PR as soon as I get information about the DZone link.