Request for minor changes
When reading the 'LATEST ON keyword' page (https://questdb.io/docs/reference/sql/latest-on/), I was confused with one sentence. So, I updated it. If you feel it helps users, pls feel free to accept the change. 

Current: When LATEST ON is provided a single column is of type SYMBOL, the query will end as soon as all distinct symbol values have been found.
Updated: When `LATEST ON` is provided a single column, and it is of type SYMBOL, the query will end as soon as all distinct symbol values are found.

I made other minor sentence updates too. Pls feel free to accept/reject the changes, as per